# FitnessArena-API/Models/traineeprogram.cs
// <auto-generated> This file has been auto generated by EF Core Power Tools. </auto-generated>
#nullable disable
using System;
using System.Collections.Generic;
using System.ComponentModel.DataAnnotations;
using System.ComponentModel.DataAnnotations.Schema;
using Microsoft.EntityFrameworkCore;

namespace FitnessArena_API.Models
{
    public partial class traineeprogram
    {
        public int? programId { get; set; }
        public int? traineeId { get; set; }
        [Key]
        public int traiProgID { get; set; }

        [ForeignKey("programId")]
        [InverseProperty("traineeprograms")]
        public virtual program program { get; set; }
        [ForeignKey("traineeId")]
        [InverseProperty("traineeprograms")]
        public virtual user trainee { get; set; }
    }
}