# FitnessArena-API/bin/Release/net6.0/swagger.json
{
  "swagger": "2.0",
  "info": {
    "title": "FitnessArena-API, Version=1.0.0.0, Culture=neutral, PublicKeyToken=null",
    "version": "1.0"
  },
  "paths": {
...
    "/api/inbodies/trainee/{id}": {
      "get": {
        "tags": [
          "inbodies"
        ],
        "produces": [
          "text/plain",
          "application/json",
          "text/json"
        ],
        "parameters": [
          {
            "in": "path",
            "name": "id",
            "required": true,
            "type": "integer",
            "format": "int32"
          }
        ],
        "responses": {
          "200": {
            "description": "Success",
            "schema": {
              "type": "array",
              "items": {
                "$ref": "#/definitions/inbody"
              }
            }
          }
        }
      }
    },
...
    "/api/programs/{id}": {
      "get": {
        "tags": [
          "programs"
        ],
        "produces": [
          "text/plain",
          "application/json",
          "text/json"
        ],
        "parameters": [
          {
            "in": "path",
            "name": "id",
            "required": true,
            "type": "integer",
            "format": "int32"
          }
        ],
        "responses": {
          "200": {
            "description": "Success",
            "schema": {
              "type": "array",
              "items": {
                "$ref": "#/definitions/program"
              }
            }
          }
        }
      },
      "put": {
        "tags": [
          "programs"
        ],
        "consumes": [
          "application/json-patch+json",
          "application/json",
          "text/json",
          "application/*+json"
        ],
        "parameters": [
          {
            "in": "path",
            "name": "id",
            "required": true,
            "type": "integer",
            "format": "int32"
          },
          {
            "in": "body",
            "name": "body",
            "schema": {
              "$ref": "#/definitions/program"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Success"
          }
        }
      },
      "delete": {
        "tags": [
          "programs"
        ],
        "parameters": [
          {
            "in": "path",
            "name": "id",
            "required": true,
            "type": "integer",
            "format": "int32"
          }
        ],
        "responses": {
          "200": {
            "description": "Success"
          }
        }
      }
    },
...
    "/api/traineeprograms": {
      "get": {
        "tags": [
          "traineeprograms"
        ],
        "produces": [
          "text/plain",
          "application/json",
          "text/json"
        ],
        "responses": {
          "200": {
            "description": "Success",
            "schema": {
              "type": "array",
              "items": {
                "$ref": "#/definitions/traineeprogram"
              }
            }
          }
        }
      },
      "post": {
        "tags": [
          "traineeprograms"
        ],
        "consumes": [
          "application/json-patch+json",
          "application/json",
          "text/json",
          "application/*+json"
        ],
        "produces": [
          "text/plain",
          "application/json",
          "text/json"
        ],
        "parameters": [
          {
            "in": "body",
            "name": "body",
            "schema": {
              "$ref": "#/definitions/traineeprogram"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Success",
            "schema": {
              "$ref": "#/definitions/traineeprogram"
            }
          }
        }
      }
    },
    "/api/traineeprograms/{id}": {
      "get": {
        "tags": [
          "traineeprograms"
        ],
        "produces": [
          "text/plain",
          "application/json",
          "text/json"
        ],
        "parameters": [
          {
            "in": "path",
            "name": "id",
            "required": true,
            "type": "integer",
            "format": "int32"
          }
        ],
        "responses": {
          "200": {
            "description": "Success",
            "schema": {
              "$ref": "#/definitions/traineeprogram"
            }
          }
        }
      },
      "put": {
        "tags": [
          "traineeprograms"
        ],
        "consumes": [
          "application/json-patch+json",
          "application/json",
          "text/json",
          "application/*+json"
        ],
        "parameters": [
          {
            "in": "path",
            "name": "id",
            "required": true,
            "type": "integer",
            "format": "int32"
          },
          {
            "in": "body",
            "name": "body",
            "schema": {
              "$ref": "#/definitions/traineeprogram"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Success"
          }
        }
      }
    },
    "/api/traineeprograms/programstrainee/{id}": {
      "get": {
        "tags": [
          "traineeprograms"
        ],
        "parameters": [
          {
            "in": "path",
            "name": "id",
            "required": true,
            "type": "integer",
            "format": "int32"
          }
        ],
        "responses": {
          "200": {
            "description": "Success"
          }
        }
      }
    },
    "/api/traineeprograms/unassign/{id}": {
      "delete": {
        "tags": [
          "traineeprograms"
        ],
        "parameters": [
          {
            "in": "path",
            "name": "id",
            "required": true,
            "type": "integer",
            "format": "int32"
          }
        ],
        "responses": {
          "200": {
            "description": "Success"
          }
        }
      }
    },
    "/api/traineeprograms/traineeprogram/{id}": {
      "get": {
        "tags": [
          "traineeprograms"
        ],
        "produces": [
          "text/plain",
          "application/json",
          "text/json"
        ],
        "parameters": [
          {
            "in": "path",
            "name": "id",
            "required": true,
            "type": "integer",
            "format": "int32"
          }
        ],
        "responses": {
          "200": {
            "description": "Success",
            "schema": {
              "type": "array",
              "items": {
                "$ref": "#/definitions/traineeprogram"
              }
            }
          }
        }
      }
    },
...
  },
  "definitions": {
...
    "inbody": {
      "type": "object",
      "properties": {
        "inbodyId": {
          "format": "int32",
          "type": "integer"
        },
        "date": {
          "format": "date-time",
          "type": "string"
        },
        "muscle": {
          "format": "double",
          "type": "number"
        },
        "minerals": {
          "format": "double",
          "type": "number"
        },
        "proteins": {
          "format": "double",
          "type": "number"
        },
        "water": {
          "format": "double",
          "type": "number"
        },
        "weight": {
          "format": "double",
          "type": "number"
        },
        "height": {
          "format": "int32",
          "type": "integer"
        },
        "traineeId": {
          "format": "int32",
          "type": "integer"
        }
      }
    },
...
    "program": {
      "type": "object",
      "properties": {
        "programId": {
          "format": "int32",
          "type": "integer"
        },
        "name": {
          "maxLength": 10,
          "minLength": 0,
          "type": "string"
        },
        "target": {
          "maxLength": 10,
          "minLength": 0,
          "type": "string"
        },
        "progDecription": {
          "maxLength": 100,
          "minLength": 0,
          "type": "string"
        },
        "coachId": {
          "format": "int32",
          "type": "integer"
        },
        "programWorkouts": {
          "type": "array",
          "items": {
            "$ref": "#/definitions/programWorkout"
          }
        }
      }
    },
    "programWorkout": {
      "type": "object",
      "properties": {
        "workoutId": {
          "format": "int32",
          "type": "integer"
        },
        "programId": {
          "format": "int32",
          "type": "integer"
        },
        "progWork": {
          "format": "int32",
          "type": "integer"
        }
      }
    },
...
    "traineeprogram": {
      "type": "object",
      "properties": {
        "programId": {
          "format": "int32",
          "type": "integer"
        },
        "traineeId": {
          "format": "int32",
          "type": "integer"
        },
        "traiProgID": {
          "format": "int32",
          "type": "integer"
        },
        "program": {
          "$ref": "#/definitions/program"
        },
        "trainee": {
          "$ref": "#/definitions/user"
        }
      }
    },
    "user": {
      "required": [
        "email",
        "fname",
        "gender",
        "lname",
        "password",
        "type"
      ],
      "type": "object",
      "properties": {
        "userId": {
          "format": "int32",
          "type": "integer"
        },
        "type": {
          "maxLength": 10,
          "minLength": 0,
          "type": "string"
        },
        "birthDate": {
          "format": "date-time",
          "type": "string"
        },
        "fname": {
          "maxLength": 20,
          "minLength": 0,
          "type": "string"
        },
        "lname": {
          "maxLength": 20,
          "minLength": 0,
          "type": "string"
        },
        "address": {
          "maxLength": 50,
          "minLength": 0,
          "type": "string"
        },
        "registerDate": {
          "format": "date-time",
          "type": "string"
        },
        "age": {
          "format": "int32",
          "type": "integer"
        },
        "gender": {
          "maxLength": 10,
          "minLength": 0,
          "type": "string"
        },
        "rating": {
          "format": "double",
          "type": "number"
        },
        "photoSrc": {
          "maxLength": 100,
          "minLength": 0,
          "type": "string"
        },
        "categoryName": {
          "maxLength": 20,
          "minLength": 0,
          "type": "string"
        },
        "target": {
          "maxLength": 20,
          "minLength": 0,
          "type": "string"
        },
        "cardNumber": {
          "format": "int32",
          "type": "integer"
        },
        "email": {
          "maxLength": 50,
          "minLength": 0,
          "type": "string"
        },
        "password": {
          "maxLength": 50,
          "minLength": 0,
          "type": "string"
        }
      }
    },
...
  }
}